remove test dependency on fixtures plugin
We had the fixtures plugin as a test dependency to test interaction with the sham controller,
but when running an app with the sham 0.2 plugin with test-app -war using grails 1.3.7
the fixtures plugin would then not be loaded. So I'm removing this as a dependency, and
people can use the plugin withot having to add manual excludes here.